drivers: mmc: omap_hsmmc: Reduce the max timeout for reset controller fsm
From OMAP3 SoCs (OMAP3, OMAP4, OMAP5, AM572x, AM571x), the DAT/CMD lines
reset procedure section in TRM suggests to first poll the SRD/SRC bit
until it is set to 0x1. But looks like that bit is never set to 1 and there
is an observable delay of 1sec everytime the driver tries to reset DAT/CMD.
(The same is observed in linux kernel).

Reduce the time the driver waits for the controller to set the SRC/SRD bits
to 1 so that there is no observable delay.

TODO: Debug why the SRC/SRD bits are never set

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>