HACK: configs: am335x_evm: Disable DM_ETH
With DM_ETH enabled, it is required that we need to use driver model
based MUSB peripheral framework to support USB RNDIS boot. This support
seems to dependant on SPL_DM and is broken in SPL atm. Proper fix
would be to fix USBETH support when SPL_DM and DM_ETH is enabled.

Signed-off-by: Vignesh R <vigneshr@ti.com>