HACK: ARM: k2g: Program PadConfig_202 before locking RSTMUX8
The PADCONFIG_202 register (0x02621328) is affected by the
locking of the RSTMUX8 register (0x02620328), and so cannot
be configured in kernel. This has been confirmed as a hardware
bug. Setup the pinmux for this pin before locking the RSTMUX8
register to allow the ICSS1 PRU1 Ethernet PHY port to work
properly.

Signed-off-by: Suman Anna <s-anna@ti.com>